Refactor: Make map_if api more eronomic
We never actually use the added capability, using a second closure which
evaluates to the correct TokenKind, gives us. So I decided to just pass
the `TokenKind` directly.

This patch also removes an unnecessary call to `next_char` aswell as a
unnecessary `expect`. Calling `next_char` can never fail because we only
land here if `peek` was successful, this means we can omit the bounds
checking done in `next_char` and just increment the cursor.