Chore: Add dependencies to Cargo.toml
Add support for the `derivative` crate which just provides yet another
way to easily generate an implementation for the `Debug` trait.
It specifically provides a way to exclude struct fields from showing up
in the Debug output. It is convinient to exlcude all the `Spans` because
they make looking at the AST really hard because of all the noise they
introduce.