Refactor: Make working with spans easier
This change is supposed to make working with spans a little bit easier. That
being said it is not unlikely that this system is going to change in the
future. Considering the number of spans compiling even a small programm
is going to involve, means that reducing the size of a `Span` is probably
going to provide a sizeable benefit in both speed and memory footprint.

We could probably get away with storing just the start and end offset into the
the source string and keep a global copy of the file path somewhere in
the AST.

But right now performance and memory footprint are smaller concerns than
getting an initial version of this language up and running. So at least
for now we're not going to think too much about these concerns :D