Feat: Add support for merging spans
Because I consider good error messages an essential feature for this
language. Tracking the location of user submitted source code on a fine
grained level is really important, which means that we track the
location of every expression including their subexpressions.

Example:

a + (b * c)
    ^-----^ Span for subexpressions
^---------^ Span for overall expression

Merging allows use to compute than `Span` of the overal expression
simply by recursively merging the subexpression spans.