Test: Seperate token and span tests
I'm not sure if seperating the tets for the generated tokens and spans
is the right coice because if we ever introduce new tokens there are now
2 places to maintain for the tests.
But the current test impl for the tokens is really clean and adding the
spans into the same test module seems less clean than seperating them.

Maybe there is a possibility for a cool macro which let's us keep the
clean code and combine the 2 tests? :D