Feat: Add rusty-hook
`rusty-hook` provides an easy way to define git hooks. These allow us to
automatically run custom commands everytime a git command is executed.
This ensures that noone forgets to run the tests before commiting a
patch for example.
Not saying that you are but if you're regularly as lost as me while coding
this is going to change your life :D