Fix: Incorrect span merging
Adding the unit tests in: 3af25ded720735d37aa75384fab1a87702ad4383 was
quite useful as they actually uncovered a bug in the way we merge spans.

Example:

Consider the 2 Spans:

Span {
    start: LineColumn {
	line: 0,
	col: 3,
    },
    end: LineColumn {
	line: 0,
	col: 5,
    },
}

Span {
    start: LineColumn {
	line: 3,
	col: 0,
    },
    end: LineColumn {
	line: 3,
	col: 2,
    },
}

The old implementation for merging these 2 spans would have looked at
the min and max of the line number:

start_line = min(0, 3) = 0
end_line   = max(0, 3) = 3

This is still correct but we get into trouble when we try to merge the
columns the same way.

start_col = min(3, 0) = 0
end_col   = max(5, 2) = 5

This results in the following span:

Span {
    start: LineColumn {
	line: 0,
	col: 0,
    },
    end: LineColumn {
	line: 3,
	col: 5,
    },
}

As you can see it is not equal to the expected span which would be:

Span {
    start: LineColumn {
	line: 0,
	col: 3,
    },
    end: LineColumn {
	line: 3,
	col: 2,
    },
}

The reason for this is simple. The first implementation fails to consider
the line number of the 2 spans when deciding between the start and end columns.
Which means if we are mering 2 spans with different line numbers where the
ending column is smaller than the starting column, the starting column
of the first line number get's set to the ending column of the second
line number. Which is not correct.

The fix for this issue is relativly simple as it just involes leveraging
the existing Cmp + PartialCmp implementation which already considers the
line number to perform the min and max comparisons between the linecolumns.