Refactor: Make span optional for internal errors
Because we use the `diagnostic` type to represent all outgoing
information to the user, it must be flexible enough to accomodate
different use cases. This includes compiler internal errors, which may
not be directly associated with a location in the users source code.

Because I think it always better to enforce invariants through the type
system rather than invalid values and conventions, I decided to make the
requirement for a valid span in the case of an internal error optional.