Refactor: Move Lit and RangeKind to lexer
While implementing the try_into conversion from token to binary
operation I needed to import the token definition from lexer to the ast
module. This resulted in a dependency cycle because the lexer depends
on Lit and RangeKind which are defined in the ast module.
As of now cargo cannot resolve such cyclic dependencies so I moved the
definition of Lit and RangeKind into the lexer module.