Refactor: Unify binary operations
Having different kinds of binary expressions in the ast does not
provided a tangable benefit as far as I can see right now. So in an
effort to simplify the AST and the parser which is responsible to
generate the AST I moved the different kind of binary expressions into a
single node.
You can use the operator to differentiate between the different kinds of
expressions now.