Feat: Add parser
The job of the parser is to take the stream of tokens provided by the
lexer and parse them into an Abstract Syntax Tree (AST).
The AST encapsulates much more information about the intend of the code
compared to just a raw stream of tokens.

Example:

main :: () {
    foo := 42;
}

By purly looking at the tokens the lexer generates for the programm
above we can make no definitive statement about the intend of the code.
All we can see for certain are some some identifiers, numbers and a bunch
of different delimiters.

By feeding the tokens throuh the parser and looking at the generated
AST we get a much more interesting picture which might look something like this:

FnDecl {
    name: "main",
    body: Block {
	VariableDefinition: {
	    name: "foo",
	    value: 42,
	}
    }
    return_type: Void,
}

As you can see we now have a much greater understanding about the source
programm.
We know the user has written a function with the name "main" which does
not return something and in this function he just creates a single
variable "foo" and then does nothing with it.
Seems like a pretty boring function if you ask me but what are supposed
to do. It's not like we can change the users source code, can we? 😉