Feat: Add cargo Makefile
`Cargo-make` allows us to easily define custom make like commands.
This provides a convenient way to automate the installation for example. In the
future we are probably going to uses this sort of Makefile quite heavily
for building the language, including the portions of this project which
are not written in rust.