Refactor: Unify checking if token is whitespace
It would be easier just to skip all whitespace characters in the lexer but I'm
really not shure if this is the right approach for this language as some
metaprogramm might need this information during compile time.

I have therefore decided to emit all whitespace and newline tokens but also
provide an easy way to check if a token is whitespace.