Feat: Add top level declarations to the ast
Our ast is comprised of different kind of nodes, where a `decl` (short
for declarations) represents the highest kind of abstraction in the ast.

Example:

main :: () {
    foo := 42 + 0;
}

In the code above `main` is a function declarations which consists of a
statement which defines the variable `foo` which intern contains the
expression `42 + 0`.