Refactor: Use new session system to report errors
Because we wan't to report as many errors as possible to the user while
running the compiler it does not make sense to abort all programm execution if
we encounter an error. Instead we aggregate them in a `Session` to dump
them all at once after all the pipelines in the compiler had their
chance to report potential errors.