Chore: Disable git test hooks
I decided to disable git testing hooks for now because we are still
activly working on the error reporting system. Which means that any
potential unit tests are quickly outdated and have to be rewritten to
work with the new error reporting system. While we are still working on
the overal architecture things like the parser it makes sense to hold of
implementing a huge number of unit tests.